% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_filename_utils.r
\name{bare_filename}
\alias{bare_filename}
\title{Return the bare filename (no path, no extension)}
\usage{
bare_filename(a_filename)
}
\arguments{
\item{a_filename}{(character) filename}
}
\value{
character filename without path and extension
}
\description{
Return the bare filename (no path, no extension)
}
\examples{

bare_filename(file.path("tmp", "test.dat"))
bare_filename(c(file.path("tmp", "test1.dat"), file.path("tmp", "test2.dat")))
bare_filename(as.list(c(
file.path("tmp", "test1.dat"), file.path("tmp", "test2.dat")
)))
}
