% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_filename_utils.r
\name{change_filename_extension}
\alias{change_filename_extension}
\title{Change filename extension.}
\usage{
change_filename_extension(a_filename, new_ext = ".tif")
}
\arguments{
\item{a_filename}{(character) filename}

\item{new_ext}{(character) new extension (with leading dot).}
}
\value{
character filename with new extension.
}
\description{
Change filename extension.
}
\examples{

change_filename_extension("test.dat")
change_filename_extension("test")
change_filename_extension("test.dat", ".txt")
change_filename_extension(c("test.txt", "test.dat"))
change_filename_extension(as.list(c("test.txt", "test.dat")))
}
