% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnamer.r
\docType{package}
\name{fnamer}
\alias{fnamer}
\alias{fnamer-package}
\title{fnamer: Collection of functions to handle and maniputate filenames.}
\description{
fnamer: Collection of functions to handle and maniputate filenames.
}
