% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_filename_utils.r
\name{get_filename_extension}
\alias{get_filename_extension}
\title{Get filename extension (with leading dot).}
\usage{
get_filename_extension(a_filename)
}
\arguments{
\item{a_filename}{filename (character).}
}
\value{
filename extension with leading dot (character).
}
\description{
Get filename extension (with leading dot).
}
\examples{
get_filename_extension("test.dat")
get_filename_extension("test")

}
