% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_filename_utils.r
\name{pks_dependencies}
\alias{pks_dependencies}
\title{Return names of all dependent packages}
\usage{
pks_dependencies(pks)
}
\arguments{
\item{pks}{Names of packages (chr)}
}
\value{
Names of all packages that depend on packages specified (chr)
}
\description{
Return names of all dependent packages
}
\examples{
pks_dependencies(c("magrittr","dplyr"))
}
