% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_filename_utils.r
\name{change_filename_extension}
\alias{change_filename_extension}
\title{Change filename extension.}
\usage{
change_filename_extension(a_filename, new_ext = ".tif")
}
\arguments{
\item{a_filename}{filename (character).}

\item{new_ext}{new extension (with leading dot) (character)}
}
\value{
filename with new extension (character)
}
\description{
Change filename extension.
}
\details{
a_filename can also be a character vector or list of character.
}
\examples{

change_filename_extension("test.dat")
change_filename_extension("test")
change_filename_extension("test.dat", ".txt")
change_filename_extension(c("test.txt", "test.dat"))
change_filename_extension(as.list(c("test.txt", "test.dat")))
}
