% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnamer.r
\docType{package}
\name{fnamer}
\alias{fnamer}
\title{fnamer: Collection of functions to handle and maniputate filenames.}
\description{
This package does not export any objects.
}
\details{
This package exports the following functions:

* \code{\link{get_filename_extension}}
* \code{\link{change_filename_extension}}
* \code{\link{bare_filename}}
}
