% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnamer.r
\docType{package}
\name{fnamer}
\alias{fnamer}
\title{fnamer: Collection of functions to handle and maniputate filenames.}
\description{
This package does not export any objects.
}
\details{
This package exports the following functions:
\itemize{
\item \code{\link{get_filename_extension}}
\item \code{\link{change_filename_extension}}
\item \code{\link{bare_filename}}
\item \code{\link{pks_dependencies}}
}
}
